+++ backend/mooc/settings.py
@@ -51,7 +51,6 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "rest_framework",
-    "rest_framework_simplejwt",
     "userprofiles",
     "corsheaders",
 ]
