+++ backend/userprofiles/views.py
@@ -1,9 +1,10 @@
 from rest_framework import generics, permissions
 from rest_framework.response import Response
 from rest_framework import status
-from .serializers import UserSerializer, GoogleAuthSerializer , LoginSerializer
+from .serializers import UserSerializer, GoogleAuthSerializer , UserLoginSerializer
 from .utils import google_authenticate
-from rest_framework.views import APIView
+from django.contrib.auth import authenticate
+from rest_framework_simplejwt.tokens import AccessToken
 
 
 class UserRegistrationApiView(generics.CreateAPIView):
@@ -21,12 +22,37 @@ class UserRegistrationApiView(generics.CreateAPIView):
         return response
     
 
-class LoginApiView(APIView):
-    def post(self, request, *args, **kwargs):
-        serializer = LoginSerializer(data=request.data)
-        if serializer.is_valid():
-            return Response(serializer.validated_data, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+class UserLoginApiView(generics.GenericAPIView):
+    serializer_class = UserLoginSerializer
+    permission_classes = [permissions.AllowAny]
+
+    def post(self, request):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        user_name = serializer.validated_data["username"]
+        password = serializer.validated_data["password"]
+        user = authenticate(username=user_name, password=password)
+        if user:
+            respObj = {
+                "status": "success",
+                "data": {
+                    "access_token": str(AccessToken.for_user(user)),
+                    "user": {
+                        "user_id": user.id,
+                        "username": user.username,
+                        "full_name": f"{user.first_name} {user.last_name}",
+                        "email": user.email,
+                    },
+                },
+            }
+            return Response(respObj, status=status.HTTP_200_OK)
+        else:
+            respObj = {
+                "status": "fail",
+                "message": ["Invalid email or password"],
+            }
+
+        return Response(respObj, status=status.HTTP_403_FORBIDDEN)
